+++ migration-library/src/main/kotlin/com/quadient/migration/service/inspirebuilder/InspireDocumentObjectBuilder.kt
@@ -1034,7 +1034,8 @@ abstract class InspireDocumentObjectBuilder(
                 val cellContentFlow = buildDocumentContentAsSingleFlow(
                     layout, variableStructure, cellModel.content, null, null, languages
                 )
-                val cellFlow = if (cellContentFlow.type === Flow.Type.SELECT_BY_INLINE_CONDITION) {
+                val cellFlow =
+                    if (cellContentFlow.type === Flow.Type.SELECT_BY_INLINE_CONDITION || cellContentFlow.type === Flow.Type.SELECT_BY_CONDITION) {
                     layout.addFlow().setType(Flow.Type.SIMPLE).setSectionFlow(true)
                         .setWebEditingType(Flow.WebEditingType.SECTION)
                         .also { it.addParagraph().addText().appendFlow(cellContentFlow) }